Refactor pandas helper functions into a dedicated module
