+++ genet/utils/dict_support.py
@@ -110,10 +110,3 @@ def combine_edge_data_lists(l1, l2):
     """
     edges = merge_complex_dictionaries({(u, v): dat for u, v, dat in l1}, {(u, v): dat for u, v, dat in l2})
     return [(u, v, dat) for (u, v), dat in edges.items()]
-
-
-def notna(value):
-    nn = pd.notna(value)
-    if isinstance(nn, ndarray):
-        return any(nn)
-    return nn
